run a socat on 0.0.0.0 to talk to "storage"
The transactor sends back a uri for storage, which for datomic-free is the same ip it thinks it's listening on.  On the transactor, this has to be 0.0.0.0 in order to be exposed.  This will cause the peer to try and communicate with 0.0.0.0:4334, which normally fails - so we connect it with socat to the correct ip.

See https://groups.google.com/forum/#!msg/datomic/wBRZNyHm03o/TNXAR4PRjY8J